move webauthn tests to a separate folder
